refactored no-op components into their own classes
