+++ cas-server-webapp/src/main/webapp/WEB-INF/spring-configuration/ticketGrantingTicketCookieGenerator.xml
@@ -45,9 +45,9 @@
     <!-- As an alternative, the cookie value verification and encryption can be turned off via
     the following configuration
 
-    <bean id="cookieCipherExecutor" class="org.jasig.cas.util.CipherExecutor.NoOpCipherExecutor" />
+    <bean id="cookieCipherExecutor" class="org.jasig.cas.util.NoOpCipherExecutor" />
 
-    <bean id="cookieValueManager" class="org.jasig.cas.web.support.CookieValueManager.NoOpCookieValueManager"
+    <bean id="cookieValueManager" class="org.jasig.cas.web.support.NoOpCookieValueManager"
           c:cipherExecutor-ref="cookieCipherExecutor" />
 
     -->
